changed integer to long

+++ src/main/java/dat3/adventureXP/repository/EquipmentStatusRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EquipmentStatusRepository extends JpaRepository<EquipmentStatus, Integer> {
+public interface EquipmentStatusRepository extends JpaRepository<EquipmentStatus, Long> {
     public Optional<EquipmentStatus> findByName(String name);
 }
